+++ src/game/phase.rs
@@ -1,0 +1,16 @@
+use std::fmt::Display;
+
+#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord, Default)]
+pub enum Phase {
+    #[default]
+    Preflop,
+    Flop,
+    Turn,
+    River,
+}
+
+impl Display for Phase {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{self:?}")
+    }
+}
